separa o arquivo index.js em app.js e seeerver.js
