+++ back-end/src/server.js
@@ -1,0 +1,4 @@
+const PORT = 3001;
+const app = require('./app');
+
+app.listen(PORT, () => console.log(`Conectado na porta ${PORT}`));
